[ 1816527 ] Eliminate OS/2-Windows type usage in portable code. Merge my changes back into trunk.  This commit removes INT and related (UINT, etc.)
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1183 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/classes/ObjectClass.cpp
@@ -798,7 +798,7 @@ RexxObject * RexxObject::processUnknown(
 /******************************************************************************/
 {
   RexxObject     *unknown_arguments[2];/* arguments to the unknown method   */
-  INT             i;                   /* loop counter                      */
+  size_t          i;                   /* loop counter                      */
   RexxArray      *argumentArray;       /* unknown method argument array     */
 
                                        /* no method for this msgname        */
